Added Annotation cross refs to SamplingFeatures & Actions.
Annotations (notes for or groups of) SamplingFeatures & Actions will be
a very common use by IEDA and others.  These needed to be explicitly
added.  I removed annotations for Variables, as I felt it was unlikely
to happen very frequently.  A DB user can always easily add the cross
reference table.