Adding RelatedActions and documentation
I created this feature branch to propose an alternative for
representing workflows and relationships among Actions.  I think it
only requires a single additional entity in the Core schema and removal
of the ParentActionID from the Actions table.  This is mainly because
there are relationships among Actions that are not strictly
parent/child relationships (particularly in sensor data workflows).  I
added a long article to the documentation in this feature branch
showing simple and more complex examples of how this can be used for
both samples and sensors.