Changes made by end of meeting
This is the last set of changes that we made during the meeting.
Changes made in this branch will be pulled out into separate feature
branches that can be merged one a time.  This may take a few days, so
we’ll keep this up and create documentation for it.