Created LabAnalysis.SpecimenBatchLines
* Removed legacy AnalysisBatchActions and BatchSampleLines tables, in
order to replace them with a potentially simpler and more effective
approach that is now supported by ODM2.
* Created LabAnalysis.SpecimenBatchLines table and added a note to
explore the following idea:
  * If an Index or Key were made from concatenating SamplingFeatureID &
ActionID from the FeatureActionResult table, then it would be possible
to track the position of a Specimen within an AnalysisBatch as it
progresses from a SpecimenPreparationAction to an
InstrumentAnalysis[=MethodType] ObservationAction. This info is tracked
in nearly every lab.
  * For a SpecimenPreparationAction, where there are no Results and
ResultID is null, such a FeatureActionIndexID would be the same as a
primary key for the FeatureActionResult table (i.e. a
FeatureActionResultID).
  * For a InstrumentAnalysis[=MethodType] ObservationAction, however,
using a FeatureActionResultID will not work, because there can by many
results for each unique Specimen-Action combo.