Moved Citations to Provenance schema
Jeff had pointed out that Citations and AuthorList should be in the
same schema, but didn’t suggest which. I agree, and I am leaning toward
moving them out of the Core and into Provenance, as I do not think they
are a required for all use cases.
* Changed the table name to AuthorLists (from AuthorList) to be
consistent with our convention (and that of OGC O&M).