Moved *DerivationEquation* field to *ResultDerivationEquations* table
For some reason I had the *ResultDerivationEquations* table be a
linking table to *DerivationEquations*, but this was completely
redundant, as a Result record can only have one DerivationEquation.  I
think I ended up with this because originally I had
DerivationEquationID as a foreign key in the Results table, but then
moved it to an optional linking table.  That was the wrong solution!
Anyways, this is right.