Manually recreated Provenance Feature Branch
Implemented the Provenance schema as it was in the previous
Provenance_FeatureBranch.  Additionally I did the following:
1. Removed DataSetCitationID in DataSets entity (it’s now redundant
given that we created the separate Citation entities
2. Checked all of the relationship names in all of the schemas to make
sure they did not have legacy names from renamed or deleted entities.
3. Deleted the InputResults entity - this was left over from before (it
just wasn’t in the diagram and is not being used)
4. Deleted ResultProvenance
5. Deleted RelatedResults entity from the ODM2Core schema and diagram
(it is now in ODM2Provenance)
6. Removed Citations from ODM2Core schema and diagram (it is now in
ODM2Provenance)
7. Added foreign key and relationship between ReferenceMaterialValues
and Citations (this was an oversight)