Added Primary Keys to all Cross-Ref Tables
As decided today and described in issue #47.
All called BridgeID.
Also added SpatialReferenceExternalIdentifiers bridge table.  * We
still want this even though we decided to rely heavily on EPSG codes
(see issue #12) because we still need to point the EPSG resolver,
right?  And we also want to allow for pointing to resolvers for local
reference frames, right?