Modified ResultValues.Aggregation... fields
* Changed the names of two fields (decided in #29):
  * AggregationInterval (or AggregationLength) would replace
AggregationDuration, to acknowledge that this concept is equally valid
for spatial coverages as it is for time series.
  * AggregationStatisticCV would replace InterpolationTypeCV, because
the former term does not properly describe the concept that it
addresses.
* Added AggregationIntervalUnitsID to the ResultValues table (decided
in #29).
I also added much better comments.
These changes had been committed to the Results_mods feature branch,
but finalizing all the changes in that branch have delayed the merge.
Merging these three decided changes now, however, will help the IEDA
team better as they work on ODM2 implementations.