Committing changes made to Equipment and Sensors
I somehow screwed up the merge of the feature branch that had all of
the changes mdd to the Equipment and Sensors schema. So, I just grabbed
the DBWrench file from the newer Sensors feature branch and replaced
the one in the Master branch.  This should be OK since there were no
commits to the DBWrench file in the Master Branch since I created the
Sensors feature branch.