Added unique constraints to code fields
Added unique constraints to all "Code" fields where the primary purpose
of the code field is to provide an alternative identifier for an entity
instance.  There are a couple of these that I did not add constraints
to (e.g., AnnotationCode and SpatialReferenceCode) because they are
nullable, and I could see use cases where they would not be unique.